+++ src/main/java/com/nhnacademy/illuwa/config/SecurityConfig.java
@@ -27,7 +27,7 @@ public class SecurityConfig {
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
         http
                 .authorizeHttpRequests(auth -> auth
-                        .requestMatchers("/", "/auth/login", "/auth/signup","/guest-login", "/verifications/**").permitAll()
+                        .requestMatchers("/", "/auth/login", "/auth/signup", "/verifications/**").permitAll()
                         .requestMatchers("/admin/**").hasRole("ADMIN")
                         .requestMatchers("/mypage").authenticated()
                         .anyRequest().permitAll()
