+++ src/test/java/com/nhnacademy/illuwa/cart/controller/CartControllerTest.java
@@ -5,7 +5,7 @@ import com.nhnacademy.illuwa.cart.dto.BookCartRequest;
 import com.nhnacademy.illuwa.cart.dto.BookCartResponse;
 import com.nhnacademy.illuwa.cart.dto.CartResponse;
 import com.nhnacademy.illuwa.cart.service.CartService;
-import com.nhnacademy.illuwa.common.controller_advice.CategoryControllerAdvice;
+import com.nhnacademy.illuwa.config.handler.CategoryControllerAdvice;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,7 +18,6 @@ import org.springframework.http.MediaType;
 import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
 import org.springframework.test.web.servlet.MockMvc;
 
-import java.awt.print.Book;
 import java.math.BigDecimal;
 import java.util.Arrays;
