+++ src/test/java/com/nhnacademy/illuwa/member/controller/AdminMemberControllerTest.java
@@ -1,8 +1,8 @@
 package com.nhnacademy.illuwa.member.controller;
 
 import com.nhnacademy.illuwa.admin.controller.AdminMemberController;
-import com.nhnacademy.illuwa.common.controller_advice.CategoryControllerAdvice;
-import com.nhnacademy.illuwa.grade.enums.GradeName;
+import com.nhnacademy.illuwa.config.handler.CategoryControllerAdvice;
+import com.nhnacademy.illuwa.member.enums.GradeName;
 import com.nhnacademy.illuwa.member.service.AdminMemberService;
 import com.nhnacademy.illuwa.common.dto.PageResponse;
 import com.nhnacademy.illuwa.member.dto.MemberResponse;
